+++ CHANGES.rst
@@ -3,6 +3,9 @@ Changelog
 
 - collective.dexteritytextindexer is integrated in Plone 6. Use plone.app.dexterity instead.
 - Index email
+- Added behavior "expert" with two new fields: organisation and region.
+  Do not use with rohberg.locationfinder. It has its own organisation (location)
+- 
 
 
 1.0.0